Updated to fix bugs.
Saving - Prompts to overwrite an exisiting script.  When a script is first opened now uses Save instead of Save As.

New Script - Prompts user where to save mission when "Build" is chosen.  A new script can not be build with out any text (to prevent building any scripts labeled Script#.script in the Gmat System directory)

Script - If the script panel has been edited and the "x" is hit in the top right corner, the user is prompted to continue (changes will be lost) or save.


git-svn-id: svn+ssh://svn.code.sf.net/p/gmat/code/trunk@3119 222415a0-0dc2-4184-858d-a27cde375526