Added "Show Mission Sequence" menu item to mission sequence node so that only mission sequence can be viewed.
Updated to read a few more options from the startup file basically for GUI testers.

PLOT_MODE = TILE
This options tiles plots when running missions so that all plots can be viewed.

DEBUG_MISSION_TREE = ON
This option writes mission sequence to the message window when command is appended or inserted to mission tree.


git-svn-id: svn+ssh://svn.code.sf.net/p/gmat/code/trunk@9929 222415a0-0dc2-4184-858d-a27cde375526

Former-commit-id: 7e43d117254855eca1b2587d7cee5e56b4294a3f