Fix for bug 2591: Moves model context deletion into ModelManager
This commit applies the patch supplied by Tristan Moody for an OpenGL crash (on 64-bit Linux) presenting as a delete-after-free segmentation fault.  There is one caveat in the patch comments:  

"As long as the ModelManager desctructor is called after all other OpenGL-related code is finished (I have not verified this), this should fix the problem."

Runs through RoutineTests are stable with this patch, so it appears that the destruction caveat is okay, but we should continue to watch it for possible issues.

git-svn-id: svn+ssh://svn.code.sf.net/p/gmat/code/trunk@10017 222415a0-0dc2-4184-858d-a27cde375526

Former-commit-id: bd01c8a8c916e3e68d05a7da7255ef0936dcacb4