Added event locators to the resource tree.
Note that there is a bug in the delete feature here.  Deleting an EventLocator from the GUI and then running without rebuilding can crash GMAT.  This bug is thought to be in the base code -- the events list in the Sandbox may be refreshing incorrectly -- but is an open issue for resolution.  It has been entered into Bugzilla as bug 2623.

git-svn-id: svn+ssh://svn.code.sf.net/p/gmat/code/trunk@10024 222415a0-0dc2-4184-858d-a27cde375526

Former-commit-id: a9a3ccc07d11f061ee2d23240e27d76f66005b8d