The rest of bug 2447 fix (file type association with GMAT)
FileManager: Added to check for startup file in the application directory when it is not found in the current working directory.
GmatApp: Used argv[0] as script file name to run when it is not a valid option. Initially implemented for Windows.
GmatMainFrame:: Updated to add script file to Scripts folder in the resource tree.


git-svn-id: svn+ssh://svn.code.sf.net/p/gmat/code/trunk@10070 222415a0-0dc2-4184-858d-a27cde375526

Former-commit-id: f983bf9253e35e52f8b6df1bc004bad1fe09e2c0