[GMT-199] Refactored the Propagate command to use the PropSetup list found in PropagationEnabledCommand.  This change lets the EventModel components enter the force model for propagation in the Propagate command.
Note that this touches the main elements of the propagation subsystem.  As such, it may have side effects that I missed during local testing.  (RoutineTests and the EventLocator folder were run back to back, and SmokeTests and the EventLocator resource folder were run in the test system.  All behave as expected.  But we should still watch things closely for a few days.)

git-svn-id: svn+ssh://svn.code.sf.net/p/gmat/code/trunk@10220 222415a0-0dc2-4184-858d-a27cde375526

Former-commit-id: 66fed7193731453cce56a3f25c8d9294bb1c8df8