Fix for following MissionTree related bugs:
GMT-63  (Reopened - Mission Tree does not handle adding elements to "If-Else" statement correctly)
Fixed the issue with appending command to nested If-Else.  

GMT-387 (Unable to insert command inside if/endif when the if is inside optimize/end optimize)
Fixed to show Append menu on control logic node inside a solver command node.  

GMT-388 (Mission Sequence Command Equationx was renamed to GMATx)
Fixed to show Equation for Assignment command and ScriptEvent for BeginScript command as it used to work.  


git-svn-id: svn+ssh://svn.code.sf.net/p/gmat/code/trunk@10259 222415a0-0dc2-4184-858d-a27cde375526

Former-commit-id: 244acfc76de38983c3c3a75f16468485b3d993bf