Fix for GMT-2559 (Transmitter, receiver, and transponder disappear from resource tree when you switch between tabs)
ResourceTree - Changed the tree item type of Transmitter, Receiver, and Transponder to GmatTree::HARDWARE from GmatTree::SENSOR in GetItemTypeAndIcon() since tree node of those is Hardware.
GuiItemManager - Added "No Antenna Selected" to antenna ComboBox in GetAntennaComboBox() so that this will be shown instead of first antenna in the list when no antenna was set to hardware.


git-svn-id: svn+ssh://svn.code.sf.net/p/gmat/code/trunk@10367 222415a0-0dc2-4184-858d-a27cde375526

Former-commit-id: ddb9337c9a4e375e697567f76dc46dee57a32bf5