Fix for GMT-2386 (Cannot zoom close to object)
OrbitViewCanvas - Changed gluPerspective() near distance from 50 to 0.1 so that it can be zoomed closer.


GMT-2735 OrbitView Code QA 

ViewCanvas, OrbitViewCanvas - Removed duplicate code of using Sun as light source. Changed the code so that if texture file does not exist it looks at the directory specified in the startup file. Added method prolog.


git-svn-id: svn+ssh://svn.code.sf.net/p/gmat/code/trunk@10606 222415a0-0dc2-4184-858d-a27cde375526

Former-commit-id: 7671b4ee208e0e5e09d3ff19ae051bd6c170e22d