Fixed plots not refreshing when moving scroll bar. Minor updates for plot persistence.
MdiChildViewFrame - Updated so that when scroll bar is moved, plots will be refreshed in OnMove().  
GmatMainFrame - Updated to return one less pixel for width and height in GetActualClientSize() to avoid showing scroll bar when plots are closed and re-run the mission.  
GmatMdiChildFrame - Updated to add one more pixel when computing position and size in Real number.  
GuiPlotReceiver - Just updated comments and debug.  




git-svn-id: svn+ssh://svn.code.sf.net/p/gmat/code/trunk@10670 222415a0-0dc2-4184-858d-a27cde375526

Former-commit-id: 87c1da96dc2149ccea757e0345cbde59d2c88ca6