WCS helped me with these:
Added two new tolerances: ZERO_MASS_TOL and ZERO_MAG_TOL for use in exception handling
Added type checking to SetRefObject() to catch the case where Primary and Secondary are not of the correct object type.  We were catching some types errors but not all, and they were caught in computation and not they are caught during initialization.
Modified exception handling to check that if mass of either primary OR secondary is too small or negative then exception is thrown. We were only throwing exception if both are too small.  However, if either mass is zero, libration point doesn't exist.
Added check for divide by zero that can occur if primary and secondary are too close together

p.s. tried calling CheckBodies() in set ref objects but that causes other problems

git-svn-id: svn+ssh://svn.code.sf.net/p/gmat/code/trunk@10699 222415a0-0dc2-4184-858d-a27cde375526

Former-commit-id: 91ef68c42f971df4398092fc75f98f93a7631d08