Fix for GMT-3042 Behavior of one-element arrays in math statements is inconsistent
Updated so that setting scalar to one element array or one element array to scalar is allowed.

Assignment - Updated to allow setting scalar to one element array or vice-versa in Validate().    
ElementWrapper - Updated to set scalar to matrix in SetValue().   
Interpreter - Updated to set scalar to matrix in SetObjectToValue().   
Array - Updated to set scalar to matrix in SetStringParameter().   
StringUtil - Added IsOneElementArray() and finished implementation of ReplaceRepeatedPlusMinusSigns().   




git-svn-id: svn+ssh://svn.code.sf.net/p/gmat/code/trunk@10907 222415a0-0dc2-4184-858d-a27cde375526

Former-commit-id: 2c9ad53397e38dd75f2a764da86cfda67c6e9d72