Fix for GMT-3042 Behavior of one-element arrays in math statements is inconsistent (complete)
Updated so that setting scalar to one element array or one element array to scalar is allowed.
This change was missed from the previous commit.

MathFunction - Updated so that if right node is not empty it will also check for valid inputs in ValidateScalarInputs().   
Atan2 - Updated to call ValidateScalarInputs() in ValidateInputs().   



git-svn-id: svn+ssh://svn.code.sf.net/p/gmat/code/trunk@10908 222415a0-0dc2-4184-858d-a27cde375526

Former-commit-id: 845d981090db62c8e4377c739fd2e4c31a7530cc