[From DJC, Thinking Systems, Inc.]
GMT-3979 and GMT-3980  Added code to handle the "All" setting for data import, and corrected the return value when the Epoch field is selected explicitly.  Also set the Epoch data to match the spacecraft that is being updated rather than the epoch read from the TVHF, correcting a coordinate system conversion reported separately. 

git-svn-id: svn+ssh://svn.code.sf.net/p/gmat/code/trunk@11957 222415a0-0dc2-4184-858d-a27cde375526

Former-commit-id: ee76ce93a4747bec54960bf1761cc5d88d66e9c1