GMT-4045 Crash after Stop with OSIRIS script
Using GCC build, I was able to duplicate the crash when I run the attached complete script with 60 sec step size. When EphemerisFiles receives the run state of End-Of-Run, it calls FinishUpWriting() and if there is not enough data it throws an exception.  I think not handling exception in Distribute() caused GMAT to crash. Added try/catch block in Distribute() and return false if exception is caught.


Former-commit-id: 05a819eaa642daf305b885c9d68fe24e02fbd671