Fix for GMT-4645, GMT-4647, GMT-4646, GMT-4652, GMT-4653, GMT-4654
GMT-4645 The Hardware missing in Show Script When Added
GMT-4647 Spacecraft information is missing in the Show Script
GMT-4646 When Rename Hardware, There are Two - Old and Renamed in the list of Spacecraft
GMT-4652 When Delete Hardware, It still stays in Spacecraft
GMT-4653 The User is able to Delete Hardware when it is used by another Object
GMT-4654 The User is able to Delete Coordinate System when it is used by another Object

Spacecraft - All above issues were due to a coding error in IsParameterReadOnly() which generated only subset of GetGeneratingString(). Fixed IsParameterReadOnly() to return correct value.

CoordPanel - Updated EnableOptions() to use epochTextCtrl->ChangeValue() instead of epochTextCtrl->Clear() which caused a crash on static link build.


Former-commit-id: 7a39aff513e837e14045094a07a2b9ad8b000a96