GMT-5160 Subscribers Behavior is NOT same when Inside Function VS. IN-LINE script (GmatFunction updates)
Updated plotting behavior inside a GmatFunction. Plots created inside a function are deleted when function run completes. Global plots remain open during the function run but plots will not be updated if drawing objects are not global. Global plots will be updated if drawing objects are global.

GmatFunction, MagageObject - Updated to make Parameter global if Parameter ref. object is global and moved it to global object store.
Function - Added a new method IsAutomaticObjectGlobal() for figuring out if Parameter ref. object is global.
FunctionManager - Wrote warning about local plots are deleted and make them global if plots to be remain open.
ParameterWrapper, TimeData, OrbitPlot, OrbitView - Just updated comments or debug output
RefData - Copied Parameter pointer in the copy constructor for GmatFunction
TimeReal - Reset initial epoch and flag unless it is global in Initialize()
ReportFile - Updated to skip data if data publishing command such as Propagate is inside a function
and the ReportFile is not a global nor a local object (i.e declared in the main script)
Subscriber - Updated not to reset status flags if subscriber is global in Initialize()
XyPlot - Updated not to delete curves if XyPlot is global in Initialize()


Former-commit-id: 116c3e06ab683ad7933e4766779ed7382fcef297