GMT-5262 Throw an exception for in-line GMAT function calls
I think fixing MathParser fixed in-line math equation with GmatFunction. So just added a flag to throw an exception if we still find it is necessary to throw an exception. This flag is turned off now.

StringUtil - Updated to check for transpose symbol in IsThereMathSymbol().
MathParser - Fixed parsing GmatFunction with math operators.
Assignment - Added a flag to throw an exception we found it is needed.


Former-commit-id: 81ed09e536177847de21af20060bf208baffa129