GMT-5361  GMT-5360
GMT-5361 FindEvents in While loop generates report for first iteration
only
GMT-5360 Overstepping ephemeris with EclipseLocator in For loop

Corrected handling of ephemFile so that old one is retained (instead of
creating a new one) and it only creates a new SpiceOrbitKernelWriter for
the new SPK, each time FindEvents is called.  This fixes FindEvents in
control flow.

Former-commit-id: 0bb4a4fcea42a014e578be479883da06b1000396