Added "earth" to atmosphere path to reflect the new location of the CSSI and Schatten files.
In a future release, we may want to look at how to set this entry, so that the "earth" piece is in the default file names instead.  That would allow central body refs toi be a bit cleaner.  Since for now the only atmospheres in the startup file are Earth, based, I think we can live with this monor blemish.


Former-commit-id: 00abe76ebf52f2ffcda841cdec413fb719aca553