GMT-5443 Ephemeris Refactoring preparation for SDO
First phase of EphemerisFile refactoring: Added the new EphemerisWriter class and EphemWriterCCSDS, EphemWriterCode500, and EphemWriterSPK to drive from it. EphemWriterWithInterpolator is an intermediate class that will handle interpolation of data points. The second phase will involve writing new util classes for CCSDS-OEM and EphemWriterCCSDS will use it to actually write out to a file. Currently EphemWriterCCSDS writes OEM data to a text file.


Former-commit-id: 0f15c467ff5c1c77aac09836bcf3f787b105d54f