GMT-5654 OrbitView Doesn't Rename Spacecraft Properly and Moved to the Wrong Field
Updated Moderator to create default FiniteBurn Parameters just for adding Parameter info into the database and to delete afterwards. Did the same for ImpulsiveBurn Parameters. The default ImpulsiveBurn object DefaultIB will no longer shown in the GUI since this object is not used in the default mission. Improved error message in RefData and ResourceTree. Updated ConfigManager and Moderator to handle exception properly from renaming an object.


Former-commit-id: 416eac967cfa4d8e40d563a765ba7f7109462936